Add logic so the admin can alias as other users when using GSI.  Also add a log message (LOG_NOTICE level) for GSI logins similar to those done for password logins.
Allowing the admin using GSI authentication to be able to alias as
other users via the clientUserName environment variable (as is
available with password authentication) was requested recently on
irod-chat as the 'Operations that require password authentication
(iCommands)' thread.  After checking into it, I found it wasn't too
difficult to implement and so went ahead and did so.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5420 ea137bd6-aa86-47b6-b90d-b101d9f19a90