Corrections for MySQL for some of the PAM-related operations (user acquiring a PAM-derived password, and iadmin rpp(remove PAM password))
Without these fixes those PAM operations would fail if the ICAT is
MySQL.

I discovered this problem after correcting a problem in the tinderbox
continuous test system so the MySQL test would run again (the Boost
runs take so long it threw off the scheme).  The 'irodsctl devtest'
(and so tinderbox) does detect this problem; MySQL issues a SQL error
so a CAT_SQL_ERR is returned when success is expected.

I'm looking at MySQL again because I'm working on integrating some
additional improvements in the iCAT (iPlant-provided) and wanted to
test them in MySQL too.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5422 ea137bd6-aa86-47b6-b90d-b101d9f19a90