Check that a user has write (or better) permission on a source collection before allowing a move out of it. This fixes the case where users could remove their home collections (and not be able to restore them).
The fix is to add another check in the chlMoveObject function in the
case of dealing with collections.  This comes into play with the 'irm
-r COLLECTION_NAME' command, where the Agent code is doing a move into
the trash collection.  A point of confusion in understanding this
problem was that with a -f ('irm -rf COLLECTION_NAME'), this problem
did not (and does not) exist as the Agent code is doing a remove
instead of a move and the checks are different.

This will also fix the case where users can also remove the
/zone/home/public collection.  Users 'own' that collection, but do not
have write access to /zone/home.

Other cases, using a direct 'imv' command for example, and working
with data-objects, is handled by somewhat differently and seems to
have been working fine.

Thanks to Michael Keller of the University of Canterbury New Zealand
for finding, reporting, and tracking this problem.  See the irod-chat
thread 'Default ACL on home directories issue' for more.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5432 ea137bd6-aa86-47b6-b90d-b101d9f19a90