Add to a script (irods2.sh) used in the tinderbox (~hourly) testing to include the major options for the run on the summary line.  This is useful in monitoring the system and confirming that particular tests are being run.
The major options are:
  PG : Reuse an existing postgres instance 
  NewPG: Download and install a new Postgres
  MySQL: Use MySQL as the ICAT DBMS instead of Postgres
  Boost: Build and test iRODS using Boost
  FUSE: Build and test FUSE
  Pound: run the Pound (load/concurrency) test
PG is used most the time but alternates with NewPG and MySQL.

These are set to run at different hours of the day so that each should
be run in a 24 hour period.

The summaries can be easily checked via the ch.pl -r command which
greps thru the tinderbox.out for the recent summary lines.

Also, change the time for the FUSE test as the NewPG run can sometimes
run long and so the hour could be missed.  

Also, update the SVN version of ch.pl with an option added a while ago
on the tinderbox host.  Add a -o option to run one-time to show the
current step/state and then exit.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5434 ea137bd6-aa86-47b6-b90d-b101d9f19a90