Disable the new gcc warning for unused-but-set-variable (now included with -Wall) as it generates more false positives than useful warnings.
Actually, for iRODS code, none of these warnings seemed to be true
positives (and a typical build had 186 of them).

The fix is to just add -Wno-unused-but-set-variable to the gcc command
line.  Hopefully, this will also be OK with older versions of gcc.

In the iRODS coding style, we'd typically call functions and hold the
return code in a variable even if not using that value later.  This
makes the code easier to read (in our opinions) and makes the value
readily available to the debugger when stepping through some
processing.  There are various cases where the caller may 'care' about
the function return code, but some when it does not.

I noticed these warnings when moving from Ubuntu 10.04 to 12.04.  The
new gcc is version 4.6.3.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5441 ea137bd6-aa86-47b6-b90d-b101d9f19a90