Add a log message for the case where an empty regParam is provided to rsModDataObjMeta and so is being ignored.
This can happen in some situations when 3.1 and 3.2 servers are mixed
due to some changes in how the packing/unpacking functions work.

The case noticed was when a 3.1 IES was being used and a 3.2 non-IES
was being installed.  As a last step in this 'irodssetup' does a iput,
iget, and diff and it was the diff that was failing.  The iput
silently failed too as this call (from agent to agent) failed and so
the size was still 0.  This does not fix that, but at least there
would something in the server log to indicate something unusual.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5457 ea137bd6-aa86-47b6-b90d-b101d9f19a90