Avoid generating PAM-derived passwords for which the scrambled form contains a single quote (') as this causes issues on some platforms (Suse 12 and perhaps others).
The logic is to just retry (generate a new password) if there is a
single quote in the scrambled form.  It will retry up to 10 times, if
needed (retries should be rare tho).  It will log the retry as a
Notice level message and if they all fail, as an Error.

This was provided in large part by the E-iRODS folks (Jason C),
thanks!


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5458 ea137bd6-aa86-47b6-b90d-b101d9f19a90