New i-command scripts, iTicket, iGet, and iDel for working with tickets and directory trees, provided by Giacomo.
These work-around some ticket limitations when working with directory
trees.  The idea is to generate a ticket for each subdirectory and use
them instead of the top-level collection ticket when needed.  See the
scripts and their help text for more.  Also see the irod-chat thread
'Problem with iticket for iget subdirectories.'  Thanks to Giacomo
Mariani of the SuperComputing Applications and Innovation Department
CINECA, Bologna, Italy, for these.

See the README.txt and its references for more.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5475 ea137bd6-aa86-47b6-b90d-b101d9f19a90