Add compatibility with OpenSSL on Centos5 (openssl pre version 1.0 in Enterprise Linux 5) as a build-time option. Thanks to by Andreas Lindqvist (Swedish NSC-SNIC) for this fix.
For those distributions that include an older version of OpenSSL,
previous to 1.0, such as Centos5, there are some compatibility issues
that we believe are safely resolved with these changes.  This is
ifdef-controlled, and is not enabled by default.

To enable this, uncomment the following line in config/config.mk:
# EL5_SSL = 1
This is below the # USE_SSL = 1 line which you uncomment to enable SSL.

Thanks to Andreas Lindqvist of the National Supercomputer Centre
(NSC), Swedish National Infrastructure for Computing (SNIC) for this
improvement.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5479 ea137bd6-aa86-47b6-b90d-b101d9f19a90