Add a test to build iRODS with gcc instead of g++.  This is run once a day via the Tinderbox system (other modes run at different hours of the day).  This is a resync with the script in use (on pivo.ucsd.edu under /tbox).
Also, changed some times for some of the tests and fixed a minor bug.

To avoid certain problems, the C complier is used on AIX, Solaris, and
whenever HPSS is used.  With this change in this tinderbox test
script, we can also test building with C on other platforms, such as
our Ubuntu test host, catching C++isms (such as // comments) if/when
they are added.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5481 ea137bd6-aa86-47b6-b90d-b101d9f19a90