Make additional checks on imeta user input for queries. This was recently reported as bugzilla item 162.
Add checks after parsing the user input to make sure that no extraeous
input was provided.  Particularly confusing would be an AND clause
which would just be ignored (if what lead up to that was OK) since it
looks for 'and'.  Other strings that the user might input and expect
to be parsed could also be ignored.

Now, after parsing the various query options, the code now checks
that the next argument is blank/null, and if not returns an
error.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5508 ea137bd6-aa86-47b6-b90d-b101d9f19a90