Significantly improve the performance of 3 SQL statements used with tickets as provided by Tony Edwin of iPlant.
Three of the SQL statements in icatMidLevelRoutines.c used in handling
tickets could cause significant memory and processor loads.  Although
the SQL will eventually only return a few rows, the cross join could
cause large numbers of records to be returned (tens of millions for
iplant) in some portion of the processing, causing the irodsAgent, for
iPlant, to consume about 10GB of memory and take around 2 minutes to
complete. When a handful of these calls would overlap in time, their
server would have a swap storm, bringing the server to a crawl.  The
problem with the cross join is subtle but in this case can be avoided
by using some sub-queries.

Again, thanks to Tony Edwin of iPlant for these SQL fixes.

I also updated the icatTicketTest.pl script, adding some additional
comments to further explain how to run it (some items I had to
re-learn to run it again).  I also modified a few of the defines to
be more convenient for current testing.

This new version passes the all the icatTicketTest.pl tests and is
somewhat faster even on small instances, especially, when one makes
the instance somewhat larger.  On Pivo (and old PC running Ubuntu
12.04.3 LTS), with 20,000 data-objects the old version of the code
(the old SQL) would take about 2m53s to run icatTicketTest.pl and the
new version takes about 2m25s.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5537 ea137bd6-aa86-47b6-b90d-b101d9f19a90