Implement logic to handle more than a preset limit (40) temporary passwords for a given user at one time.  This was an unexpected use case (these only are valid for 2 minutes are are removed when used) but was a problem for iPlant recently.
A second related problem was that once this occurred the logic to
clean up (remove all expired temporary passwords) was not triggered.
The reason for that is that the logic doesn't 'know' the password is a
temporary one until a later step and the code avoids the overhead of
the SQL that checks all password times.  The code removes all expired
temporary passwords when a temporary password succeeds.

To fix this the function now retires in this case.  It initially
operates as before (to be particularly safe), but if the first set of
passwords has been processed and if it has not yet succeeded matching
a password, it will try again, sorting the passwords by modify-time,
starting at 0, and using the last time of the last password as the
starting point if it needs to repeat.  For the case when more than the
limit (40) passwords were created in one secord, if it happens, it
will add one to the current second for the next batch (this might miss
some passwords, but will prevent an infinite loop, and it should all
get cleaned up over time).

This commit also includes additions to icatTest.pl (run as part of
'devtest') to exercise this new SQL.  Since there's a new SQL command,
there's a new rodsLog(LOG_SQL call:
if (logSQL!=0) rodsLog(LOG_SQL, "chlCheckAuth SQL 8 ");
and so the way 'devtest' operates, it must be tested to succeed.

icatTest.pl also makes sure that SQL and related logic works properly
by confirming that the last temporary password can be successfully
used.  This adds about 15 seconds to 'devtest', most of which is to
create 42 temporary passwords.  See the script for more.

The change to icatMidLeverRoutines.c is to add another argument to a
support function, an additional bind variable was needed.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5550 ea137bd6-aa86-47b6-b90d-b101d9f19a90