Changes to imeta so it will not exit with 0 when the user asks for metadata of a non-existent object.  Also, fix bugs in some cases where the code is determining if the object exists or not.
For the 5 types of objects handled, data-objects, collections,
resources, users and resource groups, imeta will now exit with a
non-success value (non 0) if the user requested listing the AVUs of an
object that does not exist.  Previously it would print either, "None"
or that it does not exist but in both cases would exit with 0.  Now,
in the 'does not exist' case, it will exit with a non-0 value.

In testing this, I noticed that there are cases where the code was not
properly differentiating between 'none' and 'does not exist'.  This
was only the cases where the user was requesting a particular
Attribute to be listed, for example, 'imeta ls -d foo attribute1', as
the code was not resetting a rcGenQuery input length value and so the
second query was not properly formed, so it could say it does not
exist when it did.  This too has been fixed, for all 5 object types.

As part of this commit I also updated icatTest.pl (part of devtest) to
check this differentiation.  It's easy now that the exit codes are
different.  I didn't need to change the test for a non-existant
attribute names ('imeta ls -d $F1 badName') because that case still
exits with 0 (the $F1 data object does exist).

This imeta change was one of a set of requests by Peter Clapham of
Sanger.  This one was easy to implement and could be considered a bug
so I was able to complete it now.  They requested this for data
objects and collecions, but I went ahead and implemented the same for
all five types.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5551 ea137bd6-aa86-47b6-b90d-b101d9f19a90