A fix for building with GSI on Mac OSX, don't try to use -z muldefs on the ld line.
The -z muldefs is not supported for the loader (ld) on OSX, so we need
to not add it when adjusting the loader line when GSI is enabled.  We
use to add -z muldefs for all but AIX OSes; so now it's for all but
AIX and OSX.

Our platform.mk.template (which is used to create platform.mk, which
is used in 'make') has been like this for GSI for some time, so I'm
not sure if no one used GSI on OSX or just worked around the problem,
or what.

Thanks to Peter Danecek of the National Institute of Geophysics and
Volcanology (INGV) of Italy for this fix.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5553 ea137bd6-aa86-47b6-b90d-b101d9f19a90