Move the scripts in server/bin/cmd to an examples subdirectory
These are primarily examples, so it is clearer to put them in an
examples directory.  The ones in server/bin/cmd can be executed via
the iexecmd command so it's better practice to have the administrator
decide which of these, if any, should be enabled.  See the new
README.txt file for more.

The previous commit, for hello, was part of this.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5558 ea137bd6-aa86-47b6-b90d-b101d9f19a90