Remove a Postgres environment variable setting for irodssetup and irodsctl.
The PGPASSWORD setting is deprecated in Postgres and no longer needed
by the iRODS installation scripts when using recent versions of
Postgres.  I've done some preliminary tests on this, and that seems to
be the case and we'll confirm all this over night when the Tinderbox
system does a full rebuild/reinstall of Postgres.

It is better not to have this in environment particularly after the
installation is completed (if the admin has left the password in
irods.config).



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5562 ea137bd6-aa86-47b6-b90d-b101d9f19a90