Change the irodssetup recommendation on setting the PATH to put the new one at the end.
This is standard practice so that more secure paths occur before less
secure ones.  If the new i-commands path is managed by someone less
trusted (non root) then it is better practice to put the i-commands
path at the end of the set of directories in the path.  This is to
avoid the possibility of the less trusted non-root person replacing a
common command like 'ls' with malware (some sort of trojan horse
program).  This is typically not a problem in iRODS as it is usually
being installed by someone with full root privileges anyway, and/or
the i-commands are moved to some other location post-build, and this
is just a recommendation being presented to the admin.  But for the
non-root case, this change presents a better recommendation for
documentation for end users.

Unchanged with this commit is that internally, irodssetup and irodsctl
place the new i-commands path at the beginning of the PATH.  But since
these are run by the irods admin, this is fine, and actually better as
this makes the new i-commands the ones that are executed, even if
there are previous versions in the PATH.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5563 ea137bd6-aa86-47b6-b90d-b101d9f19a90