Fix a double-free agent segfault when the user tries to replicate a data-object to a resource that already has a valid copy.
This was found and reported by the RENCI EiRODS folks and is tracker item
1704 in their system:
https://code.renci.org/gf/project/e-irods/tracker/?action=TrackerItemEdit&tracker_item_id=1704&tracker_id=1704
There is also an item 1705 with some related test updates.

The tracker 1704 fix is to modify an input link-list of dataObjInfo
structures in chkCopyInResc but I decided to try a different approach
and added a new function to avoid the problem more directly, by
avoiding a free of the structure a second time.  This seems like the
safer approach to me as it is more specific to the error.  It looks
like existing code, chkCopyInResc, is just checking whether this would
be a duplicate so I didn't want to possibly adversely affect any of
that, particularly since I'm not familiar with the overall logic using
that, it's somewhat involved, and seems to work OK except for the free
at the end.

The problem can be reproduced by doing a second irepl to the same
resource, at least on the platform used at DICE-UCSD using iRODS
(tracker 1704 indicates that it takes 3 copies of the data-object
there).  So for us here, it's just (assuming resc2 is not the default
resource):
  iput file1 
  irepl -R resc2 file1
  irepl -R resc2 file1

I've also added a test like the above in icatTest.pl (part of
'irodsctl devtest' and so Tinderbox).  Before the code fixes in this
commit, that test would fail as the second irepl would fail as the
agent aborted.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5571 ea137bd6-aa86-47b6-b90d-b101d9f19a90