When creating a resource, check that the resource name isn't blank.
This can be done via 'iadmin mkresc' if the user enters "" as the
resource name.  Once created, it's difficult to deal with.  We have a
number of basic checks on user input like this, but we (I) had missed
this one (probably since day one).

This was discovered via some testing done at a collaborating site and
reported to me in email.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5573 ea137bd6-aa86-47b6-b90d-b101d9f19a90