When creating a postgres database, temporarily set PGPASSWORD as it seems to be needed in some cases; follow-on to r5562 a few days ago.
This is a continuation of commit r5562 when I removed the set up of
the PGPASSWORD environment variable.  It did seem to work overnight in
tinderbox, successfully creating the datbase, but I noticed in some
testing today that in some cases createdb now prompts for the password
and fails if not entered.  The script attempts to pass it in via a
redirected file, but that no longer seems to work, and least when I
tried the irodssetup script by hand or via Tinderbox.  I'm not sure
why the same operation works in Tinderbox overnight; I tested in via
Tinderbox today too; it may be something to do with the new Postgres
build that is done at night (it might set the password somewhere
else).  But to be safer, I've added this setting and clearing of the
environment variable and it seems to work now.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5580 ea137bd6-aa86-47b6-b90d-b101d9f19a90