If PAM or GSI is enabled and iinit is creating some missing .irodsEnv info (perhaps the whole file), also prompt for & set irodsAuthScheme.
Since password is default, I made it not-prompt if the auth scheme is
the only one missing, but otherwise, if GSI and/or PAM is enabled it
will prompt for the ones available and set the authentication scheme
in the user's .irodsEnv file for them.

I did some limited testing of this in PAM mode and it seems to work
fine and should work properly in the various modes.

This was requested by Andreas Lindqvist and Giacomo on irods-chat on
the thread:'iinit not saving .irodsEnv if using other authentication
scheme than irods password'



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5581 ea137bd6-aa86-47b6-b90d-b101d9f19a90