Change rsModAccessControl to call acPostProcForModifyAccessControl only if the chlModAccessControl call succeeds.  This was recommended by Tony Edgin of iPlant on irods-chat.
Also reformatted/adjusted a little and removed some comments.  In this
case, I think it was fine to have some comments in the source
explaining a change made by a particular developer for a while, but
long term, I think it's easier to read the code if we use svn for
maintaining such information.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5585 ea137bd6-aa86-47b6-b90d-b101d9f19a90