Add a more specific error message to the user when the acCheckPasswordStrength rule does not exist.
Pete Clapham of Sanger mentioned on irods-chat that new rules like
this one can be an upgrade problem (when the acCheckPasswordStrength
rule was added), so as part making improvements related to that, I
thought this would be worthwhile.

Previously, when getting this error, iadmin would just display (two times):
Level 0 message: DEBUG: <unknown source type>
and:
ERROR: rcGeneralAdmin failed with error -1097000 NO_RULE_OR_MSI_FUNCTION_FOUND_ERR 

But now will include:
Level 1: acCheckPasswordStrength rule not found

As before, if the rule does exist but returns an error, a desription
of the error is provided to the user.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5587 ea137bd6-aa86-47b6-b90d-b101d9f19a90