Rearrange the logic so that when a SID is defined (LocalZoneSID or RemoteZoneSID) it will, by default, require that the remote server authenticate itself (also have the LocalZoneSID defined).
This makes it easier to enable server authentication, that is, you do
not need to also update a define (requireServerAuth 0) and rebuild to
make sure that the remote server authenticates itself.  This was
recommended in email discussion with Michael Keller of the U of
Canterbury.

Also see recent updates to the Server Authentication page on
irods.org.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5591 ea137bd6-aa86-47b6-b90d-b101d9f19a90