Correct a recent update for calling acPostProcForModifyAVUMetadata to test for positive values too, else it's never called for `imeta addw`.
The recently added test in rsModAVUMetadata.c was not right in all
cases since chlAddAVUMetadataWild returns the number of data objects
that were modified but we had it only calling the post-processing rule
if the return value was zero.  Error codes are always negative and
sometimes positive.

Noticed and fix provided by Dennis Roberts of iPlant. Thanks!



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5593 ea137bd6-aa86-47b6-b90d-b101d9f19a90