Correct the environment variable set for LOGFILE_INT/logfileInt in irodsctl.pl so that the interval days for log files works.
There's a 'define LOGFILE_INT "logfileInt"' which is used in the
server C code so this script needs to set environment variable
'logfileInt', not 'LOGFILE_INT'.  It looks like this was never working
properly in irodsctl.pl when implemented (r3878 mwan 2011-03-28).  The
logfileInt feature was implement long before that and so was likely
working fine, it was just the irodsctl.pl setting that failed.

Thanks to Robert Verkerk of surfsara.nl for finding, reporting, and
providing a fix for this problem.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5595 ea137bd6-aa86-47b6-b90d-b101d9f19a90