Fix an infinite loop that could occur in rare situations when using large numbers of temporary passwords.
The logic to avoid an infinite loop when there too many temporary
passwords to fit in a working array, with timestamps within a second
of each other, would fail in certain cases as the next time-stamp
value would be reset back to the previous value before doing the next
query.  Added logic to make the time-stamp update only done once per
cycle (query).

This was found and reported by the RENCI iRODS team (thanks!).  To
reproduce it at DICE-UCSD I had to modify the MAX_PASSWORDS define to
a small value (such as 2) to produce the timing conditions that would
reproduce the problem.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5596 ea137bd6-aa86-47b6-b90d-b101d9f19a90